# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movie Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/avatar/upload-url:
    post:
      operationId: getAvatarUploadUrl
      tags: [auth]
      summary: Get a signed URL for uploading a profile picture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetAvatarUploadUrlBody"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /me/avatar:
    patch:
      operationId: updateAvatar
      tags: [auth]
      summary: Update user's profile picture URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarBody"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - username
        - createdAt
    GetAvatarUploadUrlBody:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
      required:
        - filename
        - contentType
    UpdateAvatarBody:
      type: object
      properties:
        avatarUrl:
          type: string
      required:
        - avatarUrl
...
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        objectName:
          type: string
        publicUrl:
          type: string
      required:
        - uploadUrl
        - objectName
        - publicUrl
...